Move prompt resources to dedicated pkg
