+++ internal/auth0/prompt/resource.go
@@ -1,4 +1,4 @@
-package provider
+package prompt
 
 import (
 	"context"
@@ -14,7 +14,8 @@ import (
 	"github.com/auth0/terraform-provider-auth0/internal/value"
 )
 
-func newPrompt() *schema.Resource {
+// NewResource will return a new auth0_prompt resource.
+func NewResource() *schema.Resource {
 	return &schema.Resource{
 		CreateContext: createPrompt,
 		ReadContext:   readPrompt,
